Refactor table query execution
